Move populate logic to repository
